requirements.txt
pillow
markdown
django
djangorestframework
django-filter
django-simple-history
django-import-export
django-cors-headers
djangorestframework-simplejwt
django-request
tablib
tablib[xlsx]
tablib[xls]